Move rpf functions to rpf.h
